- Remoção do GreetingResource.java e GreetingResourceIT.java e seus respectivos teste - Implementação do ProdutoService.java - Implementação de Mapper de Produto para ProdutoDTO

+++ src/main/java/org/acme/dto/ProdutoDTO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 public class ProdutoDTO {
     @JsonProperty("codigoProduto")
-    private int coProduto;
+    private Long coProduto;
 
     @JsonProperty("descricaoProduto")
     private String noProduto;
